// packages/base/src/commands/ppdb-symbols/interfaces.ts
export const TYPE_PORTABLE_PDB = 'dotnet_portable_pdb'
export const VALUE_NAME_PORTABLE_PDB = 'dotnet_portable_pdb'
export const PORTABLE_PDB_FILENAME = 'dotnet_portable_pdb'

export interface MappingMetadata {
  cli_version: string
  origin_version: string
  origin: string // will contain 'datadog-ci'
  type: string

  // hex-encoded PE CodeView debug directory GUID + age, read from the build-time
  // debug ID manifest generated by dd-sdk-maui - identifies this .pdb uniquely
  debug_id: string
  assembly_name: string
  filename: string
  overwrite: boolean

  git_commit_sha?: string
  git_repository_url?: string
}